Move and clean up MultiControlledX
